chore: move bridge components/views to bridge folder
