+++ src/ui/components/bridge/ModalUnwrapBurnSummaryItems.vue
@@ -4,7 +4,7 @@
             <div
                 :key="item.key"
                 class="row"
-                :class="{ total: index === splitItems.length }"
+                :class="{ total: index === splitItems.length - 1 }"
             >
                 <div class="description">
                     {{ item.description }}
@@ -27,11 +27,13 @@
 
 <script lang="ts">
 import { computed, defineComponent, PropType, Ref, SetupContext } from "@vue/composition-api";
+import { BigNumber } from "bignumber.js";
+import { Hbar, HbarUnit } from "@hashgraph/sdk";
 
-import { formatRightPad, formatSplit } from "../../service/format";
+import { formatRightPad, formatSplit } from "../../../service/format";
 
-import { Item } from "./ModalFeeSummary.vue";
-import InfoButton from "./InfoButton.vue";
+import { Item } from "../ModalFeeSummary.vue";
+import InfoButton from "../InfoButton.vue";
 
 let KEY = 0;
 function nextItemKey(): number {
@@ -48,11 +50,55 @@ interface SplitItem {
     currency: string | null;
 }
 
+interface Amount {
+    int: string;
+    fraction: string | null;
+}
+
+interface CalculatedAmounts {
+    total: Amount;
+}
+
 export default defineComponent({
-    name: "ModalUnwrapApproveSummaryItems",
+    name: "ModalUnwrapBurnSummaryItems",
     props: { items: Array as PropType<Item[]> },
     components: { InfoButton },
     setup(props: { items: Item[] }, context: SetupContext) {
+        const calculateAmounts: Ref<CalculatedAmounts> = computed(() => {
+            let total = new BigNumber(0);
+
+            if (props.items != null) {
+                for (const item of props.items) {
+                    if (item.description === context.root.$t("interfaceUnwrapWHbar.unwrapAmount").toString()) {
+                        if (item.value instanceof Hbar) {
+                            item.value = item.value.as(HbarUnit.Hbar);
+                        }
+                        if (item.value instanceof BigNumber) {
+                            total = total.plus(item.value);
+                        }
+                    } else if (item.description === context.root.$t("interfaceWrapHbar.bridgeServiceFee").toString()) {
+                        if (item.value instanceof Hbar) {
+                            item.value = item.value.as(HbarUnit.Hbar);
+                        }
+                        if (item.value instanceof BigNumber) {
+                            total = total.minus(item.value);
+                        }
+                    }
+                }
+            }
+            const totalParts = formatSplit(total.toString());
+
+            const totalAmount: Amount = totalParts ? {
+                int: totalParts[ "int" ],
+                fraction: totalParts.fraction
+            } : {
+                int: "0",
+                fraction: "0"
+            };
+
+            return { total: totalAmount };
+        });
+
         const splitItems: Ref<readonly SplitItem[]> = computed(() => {
             // Track the long fraction part of a string
             let lengthLongestString = 0;
@@ -88,6 +134,18 @@ export default defineComponent({
                 };
             });
 
+            const computedTotal: Ref<CalculatedAmounts> = calculateAmounts; // Memoize above
+
+            // Push the the total onto the item array
+            items.push({
+                key: nextItemKey(),
+                description: "Total",
+                int: computedTotal.value.total[ "int" ],
+                fraction: computedTotal.value.total.fraction,
+                value: "",
+                currency: "ℏ"
+            });
+
             // Loop through all the items and right pad all the necessary ones
             for (const item of items) {
                 const hasFraction = item.fraction != null;
@@ -112,6 +170,7 @@ export default defineComponent({
 
         return {
             props,
+            calculateAmounts,
             splitItems
         };
     }
